refactor(startup): split tests into a separate file to bring mod.rs under the line cap
The dispatch-robustness PR (#1374) added four tests for the help-walker
and `format_clap_error`, which pushed `internal/startup/mod.rs` to
564 code lines and tripped the line-limit gate.

Move the three test modules (`help_for_tests`, `render_help_tests`,
`tests`) to `internal/startup/tests.rs` and reference them from
`mod.rs` via `mod tests;`. The production file is now 229 code
lines; the test file is 411. No semantic change to either — pure
mechanical move.
